backend/requirements.txt
fastapi
uvicorn
websockets
pillow
opencv-python
numpy
adb-shell
asyncio
python-dotenv
pydantic
requests
